backend/rag_agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools"

[project]
name = "rag-agent"
version = "0.1.0"
description = "RAG-enabled AI Agent using OpenAI Agent SDK and FastAPI"
readme = "README.md"
authors = [
    { name = "AI Robotic Book Team", email = "team@example.com" }
]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "openai>=1.3.5",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "slowapi>=0.1.8",
    "pydantic-settings>=2.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
    "pre-commit>=3.4.0",
    "httpx>=0.25.0"
]

[project.urls]
Homepage = "https://github.com/ai-robotic-book/rag-agent"
Repository = "https://github.com/ai-robotic-book/rag-agent"

[tool.setuptools.packages.find]
where = ["."]
include = ["rag_agent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra"
]

[tool.coverage.run]
source = ["rag_agent"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/.venv/*",
    "*/__pycache__/*"
]

[tool.black]
line-length = 88
target-version = ['py311']
skip-string-normalization = true

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
disallow_untyped_defs = true